move functions from interpolate.cpp to path_utilization.cpp
