+++ planning/behavior_velocity_planner/include/utilization/path_utilization.hpp
@@ -19,9 +19,13 @@
 #include <tier4_autoware_utils/geometry/geometry.hpp>
 
 #include <autoware_auto_planning_msgs/msg/path.hpp>
+#include <autoware_auto_planning_msgs/msg/path_with_lane_id.hpp>
 
 namespace behavior_velocity_planner
 {
+bool splineInterpolate(
+  const autoware_auto_planning_msgs::msg::PathWithLaneId & input, const double interval,
+  autoware_auto_planning_msgs::msg::PathWithLaneId * output, const rclcpp::Logger logger);
 autoware_auto_planning_msgs::msg::Path interpolatePath(
   const autoware_auto_planning_msgs::msg::Path & path, const double length,
   const double interval = 1.0);
@@ -29,6 +33,7 @@ autoware_auto_planning_msgs::msg::Path filterLitterPathPoint(
   const autoware_auto_planning_msgs::msg::Path & path);
 autoware_auto_planning_msgs::msg::Path filterStopPathPoint(
   const autoware_auto_planning_msgs::msg::Path & path);
+std::vector<double> calcEuclidDist(const std::vector<double> & x, const std::vector<double> & y);
 }  // namespace behavior_velocity_planner
 
 #endif  // UTILIZATION__PATH_UTILIZATION_HPP_
