+++ planning/behavior_velocity_planner/src/scene_module/intersection/scene_intersection.cpp
@@ -18,7 +18,6 @@
 #include <scene_module/intersection/scene_intersection.hpp>
 #include <scene_module/intersection/util.hpp>
 #include <utilization/boost_geometry_helper.hpp>
-#include <utilization/interpolate.hpp>
 #include <utilization/path_utilization.hpp>
 #include <utilization/util.hpp>
