+++ planning/behavior_velocity_planner/src/scene_module/no_stopping_area/scene_no_stopping_area.cpp
@@ -267,7 +267,7 @@ Polygon2d NoStoppingAreaModule::generateEgoNoStoppingAreaLanePolygon(
   const double interpolation_interval = 0.5;
   bool is_in_area = false;
   autoware_auto_planning_msgs::msg::PathWithLaneId interpolated_path;
-  if (!interpolation::splineInterpolate(
+  if (!splineInterpolate(
         path, interpolation_interval, &interpolated_path, logger_)) {
     return ego_area;
   }
