+++ planning/behavior_velocity_planner/src/utilization/path_utilization.cpp
@@ -27,6 +27,88 @@ constexpr double DOUBLE_EPSILON = 1e-6;
 
 namespace behavior_velocity_planner
 {
+bool splineInterpolate(
+  const autoware_auto_planning_msgs::msg::PathWithLaneId & input, const double interval,
+  autoware_auto_planning_msgs::msg::PathWithLaneId * output, const rclcpp::Logger logger)
+{
+  *output = input;
+
+  if (input.points.size() <= 1) {
+    RCLCPP_DEBUG(logger, "Do not interpolate because path size is 1.");
+    return false;
+  }
+
+  static constexpr double ep = 1.0e-8;
+
+  // calc arclength for path
+  std::vector<double> base_x;
+  std::vector<double> base_y;
+  std::vector<double> base_z;
+  std::vector<double> base_v;
+  for (const auto & p : input.points) {
+    base_x.push_back(p.point.pose.position.x);
+    base_y.push_back(p.point.pose.position.y);
+    base_z.push_back(p.point.pose.position.z);
+    base_v.push_back(p.point.longitudinal_velocity_mps);
+  }
+  std::vector<double> base_s = calcEuclidDist(base_x, base_y);
+
+  // remove duplicating sample points
+  {
+    size_t Ns = base_s.size();
+    size_t i = 1;
+    while (i < Ns) {
+      if (std::fabs(base_s[i - 1] - base_s[i]) < ep) {
+        base_s.erase(base_s.begin() + i);
+        base_x.erase(base_x.begin() + i);
+        base_y.erase(base_y.begin() + i);
+        base_z.erase(base_z.begin() + i);
+        base_v.erase(base_v.begin() + i);
+        Ns -= 1;
+        i -= 1;
+      }
+      ++i;
+    }
+  }
+
+  std::vector<double> resampled_s;
+  for (double d = 0.0; d < base_s.back() - ep; d += interval) {
+    resampled_s.push_back(d);
+  }
+
+  // do spline for xy
+  const std::vector<double> resampled_x = ::interpolation::slerp(base_s, base_x, resampled_s);
+  const std::vector<double> resampled_y = ::interpolation::slerp(base_s, base_y, resampled_s);
+  const std::vector<double> resampled_z = ::interpolation::slerp(base_s, base_z, resampled_s);
+  const std::vector<double> resampled_v = ::interpolation::slerp(base_s, base_v, resampled_s);
+
+  // set xy
+  output->points.clear();
+  for (size_t i = 0; i < resampled_s.size(); i++) {
+    autoware_auto_planning_msgs::msg::PathPointWithLaneId p;
+    p.point.pose.position.x = resampled_x.at(i);
+    p.point.pose.position.y = resampled_y.at(i);
+    p.point.pose.position.z = resampled_z.at(i);
+    p.point.longitudinal_velocity_mps = resampled_v.at(i);
+    output->points.push_back(p);
+  }
+
+  // set yaw
+  for (int i = 1; i < static_cast<int>(resampled_s.size()) - 1; i++) {
+    auto p = output->points.at(i - 1).point.pose.position;
+    auto n = output->points.at(i + 1).point.pose.position;
+    double yaw = std::atan2(n.y - p.y, n.x - p.x);
+    output->points.at(i).point.pose.orientation =
+      tier4_autoware_utils::createQuaternionFromYaw(yaw);
+  }
+  if (output->points.size() > 1) {
+    size_t l = output->points.size();
+    output->points.front().point.pose.orientation = output->points.at(1).point.pose.orientation;
+    output->points.back().point.pose.orientation = output->points.at(l - 2).point.pose.orientation;
+  }
+  return true;
+}
+
 autoware_auto_planning_msgs::msg::Path interpolatePath(
   const autoware_auto_planning_msgs::msg::Path & path, const double length, const double interval)
 {
@@ -178,5 +260,22 @@ autoware_auto_planning_msgs::msg::Path filterStopPathPoint(
     }
   }
   return filtered_path;
+}
+
+std::vector<double> calcEuclidDist(const std::vector<double> & x, const std::vector<double> & y)
+{
+  if (x.size() != y.size()) {
+    std::cerr << "x y vector size should be the same." << std::endl;
+  }
+
+  std::vector<double> dist_v;
+  dist_v.push_back(0.0);
+  for (unsigned int i = 0; i < x.size() - 1; ++i) {
+    const double dx = x.at(i + 1) - x.at(i);
+    const double dy = y.at(i + 1) - y.at(i);
+    dist_v.push_back(dist_v.at(i) + std::hypot(dx, dy));
+  }
+
+  return dist_v;
 }
 }  // namespace behavior_velocity_planner
